+++ node_helper.js
@@ -85,7 +85,7 @@ module.exports = NodeHelper.create({
         })
       }
       else {
-        self.sendSocketNotification('MMM-MYSCOREBOARD-SCORE-UPDATE', { instanceId: payload.instanceId, index: payload.league, scores: {} })
+        self.sendSocketNotification('MMM-MYSCOREBOARD-SCORE-UPDATE', { instanceId: payload.instanceId, index: payload.league, scores: [] })
       }
       if (payload.whichDay.yesterday === 'yes') {
         provider2.getScores(payload, moment(payload.gameDate).subtract(1, 'day'), function (scores) {
@@ -93,7 +93,7 @@ module.exports = NodeHelper.create({
         })
       }
       else if (payload.whichDay.yesterday === 'erase') {
-        self.sendSocketNotification('MMM-MYSCOREBOARD-SCORE-UPDATE-YD', { instanceId: payload.instanceId, index: payload.league, scores: {} })
+        self.sendSocketNotification('MMM-MYSCOREBOARD-SCORE-UPDATE-YD', { instanceId: payload.instanceId, index: payload.league, scores: [] })
       }
     }
     else if (notification == 'MMM-MYSCOREBOARD-GET-LOCAL-LOGOS') {