+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.8.1](https://github.com/dathbe/MMM-MyScoreboard/compare/v4.8.0...v4.8.1) - 2025-04-17
+
+- BUG FIX: Scores would disappear when a league with no games was before a league with games in config
+- Some new channel logos
+
 ## [4.8.0](https://github.com/dathbe/MMM-MyScoreboard/compare/v4.7.1...v4.8.0) - 2025-04-16
 
 - **NEW FEATURE**: Added `localMarkets` config option to display broadcasts from your local markets